Little renaming Moved enrichment into `CONFIG.TextEditor.enrichers` as per guidance from https://discord.com/channels/170995199584108546/722559135371231352/1192834854614737068
