+++ src/scripts/hooks/index.js
@@ -7,7 +7,7 @@ import { Ready } from "./ready.js";
 import { DeleteToken } from "./tokenDocumentDeleted.js";
 import { AutocompleteInlinePropertiesSetup } from "./aip-setup.js";
 import { GetSceneControlButtons } from "./get-scene-control-buttons.js";
-import { PimpJournal } from "./pimp-journal.js";
+import { CompendiumBrowserInlineEnricher } from "./compendium-browser-inline-enricher.js";
 import { TagifySheets } from "./tagify-sheets.js";
 
 export const PtuHooks = {
@@ -23,7 +23,7 @@ export const PtuHooks = {
             DeleteToken,
             AutocompleteInlinePropertiesSetup,
             GetSceneControlButtons,
-            PimpJournal,
+            CompendiumBrowserInlineEnricher,
             TagifySheets
         ]
         for(const listener of listeners) listener.listen();
